@@ -4,29 +4,32 @@ import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
-import org.joda.time.DateTime;
-import org.joda.time.format.DateTimeFormat;
-import org.joda.time.format.DateTimeFormatter;
 
 import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 
-public class CustomDateTimeSerializer extends StdSerializer<DateTime> {
+public class CustomDateTimeSerializer extends StdSerializer<LocalDateTime> {
 
-    private static DateTimeFormatter formatter =
-            DateTimeFormat.forPattern("yyyy-MM-dd HH:mm");
+    /**
+     * 
+     */
+    private static final long serialVersionUID = 1L;
 
+    private static DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
+    
     public CustomDateTimeSerializer() {
         this(null);
     }
 
-    public CustomDateTimeSerializer(Class<DateTime> t) {
+    public CustomDateTimeSerializer(Class<LocalDateTime> t) {
         super(t);
     }
 
     @Override
     public void serialize
-            (DateTime value, JsonGenerator gen, SerializerProvider arg2)
+            (LocalDateTime value, JsonGenerator gen, SerializerProvider arg2)
             throws IOException, JsonProcessingException {
-        gen.writeString(formatter.print(value));
+        gen.writeString(value.format(formatter));
     }
 }
