@@ -1,7 +1,6 @@
 package org.ozwillo.dcexporter.service;
 
 import javaslang.control.Either;
-import org.joda.time.DateTime;
 import org.ozwillo.dcexporter.dao.DcModelMappingRepository;
 import org.ozwillo.dcexporter.dao.SynchronizerAuditLogRepository;
 import org.ozwillo.dcexporter.model.Ckan.CkanResource;
@@ -16,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.time.LocalDateTime;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -37,8 +37,12 @@ public class DcModelMappingService {
     private String ckanUrl;
 
 
-    public DcModelMapping getById(String id) {
-        return dcModelMappingRepository.findById(id);
+    public Either<String, DcModelMapping> getById(String id) {
+        Optional<DcModelMapping> opt = dcModelMappingRepository.findById(id);
+        if (opt.isPresent())
+            return Either.right(opt.get());
+        else
+            return Either.left("dataset.notif.not_exist");
     }
 
     public Either<String, DcModelMapping> add(DcModelMapping dcModelMapping) {
@@ -65,7 +69,7 @@ public class DcModelMappingService {
         dcModelMapping.setOrganizationId(ckanDataset.getOrganization().getId());
 
         dcModelMapping = dcModelMappingRepository.save(dcModelMapping);
-        SynchronizerAuditLog newAuditLog = new SynchronizerAuditLog(dcModelMapping.getType(), SynchronizerStatus.PENDING, null,  DateTime.now());
+        SynchronizerAuditLog newAuditLog = new SynchronizerAuditLog(dcModelMapping.getType(), SynchronizerStatus.PENDING, null,  LocalDateTime.now());
         synchronizerAuditLogRepository.save(newAuditLog);
         return Either.right(dcModelMapping);
     }
@@ -76,7 +80,7 @@ public class DcModelMappingService {
 
         if(oldDcModelMapping == null) return Either.left("dataset.notif.not_exist");
 
-        SynchronizerAuditLog newAuditLog = new SynchronizerAuditLog(dcModelMapping.getType(), SynchronizerStatus.MODIFIED, null,  DateTime.now());
+        SynchronizerAuditLog newAuditLog = new SynchronizerAuditLog(dcModelMapping.getType(), SynchronizerStatus.MODIFIED, null,  LocalDateTime.now());
         synchronizerAuditLogRepository.save(newAuditLog);
         dcModelMappingRepository.save(dcModelMapping);
 
@@ -127,14 +131,23 @@ public class DcModelMappingService {
     }
 
     public Either<String, DcModelMapping> deleteById(String id) {
-        DcModelMapping dcModelMapping = dcModelMappingRepository.findById(id);
-        if ( dcModelMapping == null ) return Either.left("dataset.notif.not_exist");
-        else if ( dcModelMapping.isDeleted() ) return Either.left("dataset.notif.not_synchronized");
-        dcModelMapping.getCkanResourceId().forEach((key,resourceId) -> {
-            ckanService.deleteResource(resourceId);
-        });
-        dcModelMapping.setDeleted(true);
-        dcModelMappingRepository.save(dcModelMapping);
-        return Either.right(dcModelMapping);
+        Optional<DcModelMapping> opt = dcModelMappingRepository.findById(id);
+        
+        if (opt.isPresent()) {
+            DcModelMapping dcModelMapping = opt.get();
+            if (dcModelMapping.isDeleted()) 
+                return Either.left("dataset.notif.not_synchronized");
+            
+            dcModelMapping.getCkanResourceId().forEach((key,resourceId) -> {
+                ckanService.deleteResource(resourceId);
+            });
+            
+            dcModelMapping.setDeleted(true);
+            dcModelMappingRepository.save(dcModelMapping);
+            return Either.right(dcModelMapping);
+        }
+        else {
+            return Either.left("dataset.notif.not_exist");
+        }
     }
 }
